pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.4",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "doraemon"
version = "0.2.0"
description = "Python doraemon dev tool box"
authors = [ "Vincent Duan <vincent.duan95@outlook.com>",]
maintainers = [ "Vincent Duan <vincent.duan95@outlook.com>",]
repository = "https://github.com/Duan-JM/doraemon"
classifiers = [ "Development Status :: 5 - Production/Stable", "Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Topic :: Software Development :: Libraries",]
keywords = [ "tools"]
include = [ "LICENSE", "README.md"]
readme = "README.md"
license = "Apache-2.0"

[[tool.poetry.source]]
name = "aliyun for china users"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[tool.poetry.dependencies]
python = ">3.9,<3.11"
structlog = "24.1.0"
structlog_sentry = ">=2.0.3"
openai = "^1.37"
numbers_parser = ">=4.13.2"
pandas = "==2.2.3"
dacite = "^1.8.1"
aiohttp = "^3.10.10"
asyncio = "^3.4.3"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
opentelemetry-api = "==1.33.1"
opentelemetry-sdk = "==1.33.1"
opentelemetry-semantic-conventions = "==0.54b1"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-exporter-otlp-proto-http = "^1.24.0"
pyyaml = "==5.3.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.23.5"
pytest-xdist = "^3.2.1"
pytest-mock = ">0"
pytest = "^7.1.3"
black = "^22.10.0"
tqdm = "^4.66.4"
ipdb = "^0.13.13"

[tool.pytest.ini_options]
pythonpath = "src"
log_cli = true
log_cli_level = "DEBUG"
markers = [
    "need_middlewares: marks tests as need external middlewares such as kafka",
]

[tool.coverage.run]
source = ["src"]
omit = []
